code/requirements.txt
base64==1.7.1
requests==2.26.0
python-dotenv==0.19.1
